+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/CoinBalanceItem.kt
@@ -1,0 +1,11 @@
+package io.horizontalsystems.bankwallet.modules.multiswap
+
+import io.horizontalsystems.bankwallet.entities.CurrencyValue
+import io.horizontalsystems.marketkit.models.Token
+import java.math.BigDecimal
+
+data class CoinBalanceItem(
+    val token: Token,
+    val balance: BigDecimal?,
+    val fiatBalanceValue: CurrencyValue?,
+)
