+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/PriceImpactLevel.kt
@@ -1,0 +1,9 @@
+package io.horizontalsystems.bankwallet.modules.multiswap
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+enum class PriceImpactLevel : Parcelable {
+    Negligible, Normal, Warning, Forbidden
+}
