+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/SwapError.kt
@@ -1,0 +1,5 @@
+package io.horizontalsystems.bankwallet.modules.multiswap
+
+sealed class SwapError : Throwable() {
+    object InsufficientBalanceFrom : SwapError()
+}
