+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/SwapSelectCoinViewModel.kt
@@ -16,7 +16,6 @@ import io.horizontalsystems.bankwallet.core.supported
 import io.horizontalsystems.bankwallet.core.supports
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.CurrencyValue
-import io.horizontalsystems.bankwallet.modules.multiswap.SwapMainModule.CoinBalanceItem
 import io.horizontalsystems.bankwallet.modules.receive.FullCoinsProvider
 import io.horizontalsystems.marketkit.models.BlockchainType
 import io.horizontalsystems.marketkit.models.Token
