+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/SwapViewModel.kt
@@ -228,14 +228,14 @@ data class SwapUiState(
     val error: Throwable?,
     val availableBalance: BigDecimal?,
     val priceImpact: BigDecimal?,
-    val priceImpactLevel: SwapMainModule.PriceImpactLevel?,
+    val priceImpactLevel: PriceImpactLevel?,
     val priceImpactCaution: HSCaution?,
     val fiatAmountIn: BigDecimal?,
     val fiatAmountOut: BigDecimal?,
     val fiatPriceImpact: BigDecimal?,
     val currency: Currency,
     val fiatAmountInputEnabled: Boolean,
-    val fiatPriceImpactLevel: SwapMainModule.PriceImpactLevel?,
+    val fiatPriceImpactLevel: PriceImpactLevel?,
     val timeRemaining: Long?,
     val timeout: Boolean,
     val timeRemainingProgress: Float?
